fix/mypage module
