+++ src/screens/primary/primary-my-page/module/mypage-content/index.ts
@@ -1,0 +1,1 @@
+export * from './mypage-content.module';
