--- src/screens/primary/primary-my-page/module/mypage-footer/index.ts
@@ -1,1 +1,0 @@
-export * from './mypage-footer.module';
